Add 突然の and move some functions to grim
