+++ src/eastasianwidth.d.ts
@@ -1,0 +1,3 @@
+declare module 'eastasianwidth' {
+  function length(text: string): number;
+}
